+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 2025-10-13
 
+
+- Remove explicit cache-related options from RocksDB configuration and reverted optimistic transactions to reduce RAM usage [#4853](https://github.com/lambdaclass/ethrex/pull/4853)
 - Remove unnecesary mul in ecpairing [#4843](https://github.com/lambdaclass/ethrex/pull/4843)
 
 ### 2025-10-06